% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_current_tutorial.R
\name{check_current_tutorial}
\alias{check_current_tutorial}
\title{Check current tutorial}
\usage{
check_current_tutorial()
}
\description{
An add-in for formatting tutorials.

Uses format_tutorial() to format the tutorial Rmd open
in the current editor
}
