% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determine_code_chunk_name.R
\name{determine_code_chunk_name}
\alias{determine_code_chunk_name}
\title{Determine the code chunk name of a new exercise in a tutorial.}
\usage{
determine_code_chunk_name(file_path = NULL)
}
\arguments{
\item{file_path}{Character string of the file path to the tutorial}
}
\value{
The section id of the exercise based on its section
}
\description{
Determine the code chunk name of a new exercise in a tutorial.
}
