% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determine_exercise_number.R
\name{determine_exercise_number}
\alias{determine_exercise_number}
\title{Finds the number of the next exercise in a tutorial}
\usage{
determine_exercise_number(file_path = NULL)
}
\arguments{
\item{file_path}{Character string of the file path to the tutorial}
}
\value{
The next exercise number based on the file argument or the active document.
}
\description{
Finds the number of the next exercise in a tutorial
}
