% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.check_quarto}
\alias{.check_quarto}
\title{Package build dependencies}
\usage{
.check_quarto()
}
\description{
This function ensures renv detects quarto as a dependency
while keeping it in Suggests rather than Imports.
}
\keyword{internal}
