% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_tutorial.R
\name{format_tutorial}
\alias{format_tutorial}
\title{Format Tutorial}
\usage{
format_tutorial(file_path)
}
\arguments{
\item{file_path}{Character string}
}
\value{
formatted document with correct code and hint chunk labels
}
\description{
A function for formatting tutorial Rmd files.

It adjusts code chunk labels correctly.

This should make things easier for tutorial writers because it will allow
fast chunk label updating and quick formatting and checking.

There is also no need to keep track of the exercise numbers because it is
done for you in the addin.
}
