% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_submissions_from_learnr_session.R
\name{get_submissions_from_learnr_session}
\alias{get_submissions_from_learnr_session}
\title{Get Submissions from a learnr Session}
\usage{
get_submissions_from_learnr_session(sess)
}
\arguments{
\item{sess}{session object from shiny with learnr}
}
\value{
a list which includes the exercise submissions of tutorial
}
\description{
Get Submissions from a learnr Session
}
\examples{
\dontrun{
  get_submissions_from_learnr_session(session)
}

}
