% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_tutorials.R
\name{knit_tutorials}
\alias{knit_tutorials}
\title{Knit all the tutorials in a package}
\usage{
knit_tutorials(package)
}
\arguments{
\item{package}{character vector of the package name to be tested}
}
\description{
Knit all the tutorials in a package
}
