% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_exercise.R
\name{make_exercise}
\alias{make_exercise}
\title{Make Exercise}
\usage{
make_exercise(type = "code")
}
\arguments{
\item{type}{character of question type. Must be one
of c("code", "no-answer", "yes-answer")}
}
\value{
exercise skeleton with appropriate labels and numbers
}
\description{
An add-in for writing tutorials.

It reads the latest exercise and section name
and then adds an exercise skeleton.

This should make things easier for tutorial-writers
because now a fast click can create most of the exercise for you.

There is also no need to keep track of the exercise numbers
because it is done for you in the add-in.
}
