% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_exercise.R
\name{make_no_answer}
\alias{make_no_answer}
\title{Make Exercise without an Answer}
\usage{
make_no_answer()
}
\description{
It appears that the RStudio addins must have function names only as the
Binding value. In other words, you can't have make_exercise(type =
'no-answer') as the value. So, we need a function which makes this call.
}
