% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_yes_answer.R
\name{make_yes_answer}
\alias{make_yes_answer}
\title{Make Exercise witn an Answer}
\usage{
make_yes_answer()
}
\description{
It appears that the RStudio addins must have function names only as the
Binding value. In other words, you can't have make_exercise(type =
'yes-answer') as the value. So, we need a function which makes this call.
}
