% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_tutorial_paths.R
\name{return_tutorial_paths}
\alias{return_tutorial_paths}
\title{Return all the paths to the tutorials in a package}
\usage{
return_tutorial_paths(package)
}
\arguments{
\item{package}{Character vector of the package name to be tested.}
}
\value{
Character vector of the full paths to all installed tutorials in
\code{package}.
}
\description{
Takes a package name and returns a character vector of all the
paths to tutorials in the installed package. Assumes that every Rmd file in
inst/tutorials/*/ is a tutorial, which should be true.
}
\examples{
return_tutorial_paths('learnr')
  
}
