% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_binary_only_in_r_profile.R
\name{set_binary_only_in_r_profile}
\alias{set_binary_only_in_r_profile}
\title{Set pkgType to binary in .Rprofile}
\usage{
set_binary_only_in_r_profile()
}
\value{
No return value, called for side effects.
}
\description{
This functions sets the \code{pkgType} global option to "binary" in your
\code{.Rprofile}. New R users, especially those on Windows, should never install
from source. Doing so fails too often, and too confusingly. It also sets the
value for this R session. So, you do not need to either restart R nor source
the .Rprofile by hand.

You can examine your \code{.Rprofile} to confirm this change with
usethis::edit_r_profile()
}
