% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submission_functions.R
\docType{data}
\name{submission_server}
\alias{submission_server}
\alias{submission_ui}
\title{Tutorial submission functions}
\format{
An object of class \code{shiny.tag} of length 3.
}
\usage{
submission_server(session)

submission_ui
}
\arguments{
\item{session}{session object from shiny with learnr}
}
\value{
an html div element
}
\description{
The following function was modified from Colin Rundel's learnrhash package,
available at https://github.com/rundel/learnrhash.

Note that when including these functions in a learnr Rmd document it is
necessary that the server function, \code{submission_server()}, be included in an
R chunk where \code{context="server"}.
}
\examples{
\dontrun{
  submission_server()
}

\dontrun{
  submission_ui
}

}
\keyword{datasets}
