% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_knit_tutorials.R
\name{test_knit_tutorials}
\alias{test_knit_tutorials}
\title{Knit all the tutorials in a package}
\usage{
test_knit_tutorials(package)
}
\arguments{
\item{package}{character vector of the package name to be tested}
}
\description{
Knit all the tutorials in a package
}
