% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_answers.R
\name{write_answers}
\alias{write_answers}
\title{Write tutorial answers to file}
\usage{
write_answers(file, session)
}
\arguments{
\item{file}{Path to write HTML answers file.}

\item{session}{Session object from \code{Shiny} with \code{learnr}, or a list of submission objects (for testing).}
}
\value{
NULL (side effect: writes file)
}
\description{
Take a tutorial session, extract all submitted answers,
and write out an HTML file with those answers.
}
\details{
Currently only records questions/exercises the student has completed.
It may be better to include all questions/exercises, with unanswered as NA.

Data structure of a learnr submission object:
- \code{obj$data$answer[[1]]} question answer
- \code{obj$data$code[[1]]} exercise answer
- \code{obj$type[[1]]} "exercise_submission"
- \code{obj$id[[1]]} "id"
}
\examples{
\dontrun{
  write_answers("getting-started_answers.html", sess)
}

}
