% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_current_tutorial.R
\name{check_current_tutorial}
\alias{check_current_tutorial}
\title{Check current tutorial}
\usage{
check_current_tutorial()
}
\description{
A function which formats the exercise numbers and code chunk
labels correctly. \code{check_current_tutorial()} reads in the tutorial
which is in the active editor window in Positron. It determines what
the number of each exercise should be and fixes any mistakes. It
ensures that all code chunk labels are the correct function of the
section title and exercise number.
}
