% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_tutorial_defaults.R
\name{check_tutorial_defaults}
\alias{check_tutorial_defaults}
\title{Check that all the tutorials in a package have standard components}
\usage{
check_tutorial_defaults(package)
}
\arguments{
\item{package}{character vector of the package name to be tested}
}
\description{
Check that all the tutorials in a package have standard components
}
