% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_tutorial_defaults.R
\name{check_tutorial_defaults}
\alias{check_tutorial_defaults}
\title{Check that a tutorial has the standard components}
\usage{
check_tutorial_defaults(tutorial_paths)
}
\arguments{
\item{tutorial_paths}{character vector of the paths to the tutorials to be tested}
}
\description{
Check that a tutorial has the standard components
}
\examples{
\dontrun{
  check_tutorial_defaults('/path/to/tutorial.Rmd')
}

}
