% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_submissions_from_learnr_session.R
\name{get_submissions_from_learnr_session}
\alias{get_submissions_from_learnr_session}
\title{Return a list of tutorial answers}
\usage{
get_submissions_from_learnr_session(sess)
}
\arguments{
\item{sess}{session object from shiny with learnr}
}
\value{
a list which includes the exercise submissions of tutorial
}
\description{
Grabs information from the \code{learnr} session environment, not
directly from the session object itself. Since we are using the session
environment, we currently don't (?) have a way to save the environment and
hence can't test this function.
}
