% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_tutorials.R
\name{knit_tutorials}
\alias{knit_tutorials}
\title{Knit all the tutorials in a package}
\usage{
knit_tutorials(tutorial_paths)
}
\arguments{
\item{tutorial_paths}{character vector of the paths to the tutorials to be
knitted}
}
\description{
Knit all the tutorials in a package
}
\examples{
\dontrun{
knit_tutorials(tutorial_paths = c("/path/to/this/tutorial.Rmd",
                                  "/path/to/that/tutorial.Rmd"))
}

}
