% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_tutorials.R
\name{knit_tutorials}
\alias{knit_tutorials}
\title{Knit a set of tutorials}
\usage{
knit_tutorials(tutorial_paths)
}
\arguments{
\item{tutorial_paths}{Character vector of the paths to the tutorials to be
knitted.}
}
\value{
No return value, called for side effects.
}
\description{
We define "testing" a tutorial as (successfully) running
\code{render()} on it. This function renders all the tutorials provided in
\code{tutorial_paths}. There is no check to see if the rendered file looks OK.
If a tutorial fails to render, then an error will be generated which will
propagate to the caller.
}
\examples{
\dontrun{
  knit_tutorials(tutorial_paths = return_tutorial_paths("tutorial.helpers"))
}

}
