% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_submissions.R
\name{process_submissions}
\alias{process_submissions}
\title{Process submissions from HTML files}
\usage{
process_submissions(path, pattern, return_value = "Summary")
}
\arguments{
\item{path}{A character string specifying the path to the directory containing the submission files.}

\item{pattern}{A character string representing the regular expression pattern to match against the filenames.}

\item{return_value}{A character string specifying the desired return value. "All" returns the full submission data,
and "Summary" (default) returns a summary with one row per processed file.}
}
\value{
A tibble containing either the full submission data or a summary based on the specified return value.
\itemize{
\item If return_value is "All", the tibble has the following columns:
\itemize{
\item id: The unique identifier for each submission.
\item submission_type: The type of submission.
\item answer: The answer or response corresponding to each submission.
}
\item If return_value is "Summary", the tibble has the following columns:
\itemize{
\item name: The value in the "information-name" column.
\item email: The value in the "information-email" column.
\item time: The value in the "download-answers" column.
\item answers: The number of rows in the submitted file.
}
}
}
\description{
This function takes a directory path and a regular expression pattern as input and processes
the submission files in the specified directory whose filenames have the suffix ".html"
and match the given pattern. It extracts the relevant information from the HTML tables
and returns either the full submission data or a summary based on the specified return value.
}
\examples{
\dontrun{
process_submissions("path/to/submissions", "project", return_value = "All")
}

}
