% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_rstudio_settings.R
\name{set_rstudio_settings}
\alias{set_rstudio_settings}
\title{Select smart setting for RStudio}
\usage{
set_rstudio_settings()
}
\value{
No return value, called for side effects.
}
\description{
This functions selects RStudio settings which make learning easier for new
users. These settings are stored in: ~/.config/rstudio/rstudio-prefs.json.
The most important changes are \code{save_workspace} to \code{"never"},
\code{load_workspace} to \code{FALSE}, and \code{"insert_native_pipe_operator"} to \code{TRUE}.
}
