% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_answers.R
\name{write_answers}
\alias{write_answers}
\title{Write tutorial answers to file}
\usage{
write_answers(file, session)
}
\arguments{
\item{file}{Path to write HTML answers file.}

\item{session}{Session object from \code{Shiny} with \code{learnr}, or a list of submission objects (for testing).}
}
\value{
NULL (side effect: writes file)
}
\description{
Take a tutorial session, extract all submitted answers,
and write out an HTML file with those answers.
}
\details{
Currently only records questions/exercises the student has completed.
It may be better to include all questions/exercises, with unanswered as NA.
See also: https://mastering-shiny.org/action-transfer.html#downloading-reports

Data structure of a learnr submission object:
- obj$data$answer[\link{1}] question answer
- obj$data$code[\link{1}] exercise answer
- obj$type[\link{1}] "exercise_submission"
- obj$id[\link{1}] "id"
}
\examples{
if(interactive()){
  write_answers("getting-started_answers.html", sess)
}

}
