% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_answers.R
\name{write_answers}
\alias{write_answers}
\title{Write Tutorial Answers}
\usage{
write_answers(file, session, is_test = FALSE)
}
\arguments{
\item{file}{location to render answers to. Output file type determined by
file suffix.}

\item{session}{session object from shiny with learnr}

\item{is_test}{check if testing function}
}
\description{
Write Tutorial Answers
}
