% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_answers.R
\name{write_answers}
\alias{write_answers}
\title{Write tutorial answers to file}
\usage{
write_answers(file, obj)
}
\arguments{
\item{file}{Output file path (should end in .html).}

\item{obj}{Either a Shiny session object (from learnr) or a list of submissions
(as returned by get_submissions_from_learnr_session()).}
}
\description{
Take a tutorial session (or a submission list), extract all submitted answers,
and write out an HTML file with those answers.
}
